refactor: LottoService 로직을 다른 클래스로 모두 이동
