+++ src/main/java/lotto/domain/LottoService.java
@@ -13,19 +13,5 @@ public class LottoService {
         return lottoRepository;
     }
 
-    public double calculateEarningRate(final RatingInfo ratingInfo, final long money) {
-        return totalSum(ratingInfo) / money;
-    }
-
-    private double totalSum(final RatingInfo ratingInfo) {
-        double sum = 0;
-        for (Rating rating : Rating.values()) {
-            if (rating == Rating.MISS) {
-                break;
-            }
-            sum += rating.getReward() * ratingInfo.get(rating);
-        }
-        return sum;
-    }
 }
